refactor: extract shared CLI utilities and revert version bump
- Revert package versions from 0.0.104 back to 0.0.103
- Extract isCommandAvailable/detectRuntimes to src/cli/lib/runtimes.ts
- Extract activateAndSave flow to src/cli/lib/activate.ts
- Refactor register.ts and login.ts to use shared modules
- Update test mocks for new module structure
