+++ src/app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alook/app",
-  "version": "0.0.104",
+  "version": "0.0.103",
   "description": "Run Alook locally — one command, no clone needed",
   "license": "Apache-2.0",
   "homepage": "https://github.com/alookai/alook#readme",
