+++ src/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alook/cli",
-  "version": "0.0.104",
+  "version": "0.0.103",
   "description": "Alook CLI — Enable Your Person Colleague",
   "license": "Apache-2.0",
   "homepage": "https://github.com/alookai/alook#readme",
