+++ src/shared/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alook/shared",
-  "version": "0.0.104",
+  "version": "0.0.103",
   "private": true,
   "exports": {
     ".": "./src/index.ts",
