+++ src/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alook/web",
-  "version": "0.0.104",
+  "version": "0.0.103",
   "private": true,
   "scripts": {
     "dev": "next dev",
