+++ src/ws-do/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@alook/ws-do",
-  "version": "0.0.104",
+  "version": "0.0.103",
   "private": true,
   "scripts": {
     "dev": "wrangler dev --persist-to ../web/.wrangler/state",
